requirements.txt
comet_ml==3.47.2
datasets==3.2.0
einops==0.8.0
faiss_cpu==1.10.0
matplotlib==3.10.0
mtcnn==1.0.0
numpy==2.2.2
opencv_python==4.10.0.84
Pillow==11.1.0
pytorch_lightning==2.4.0
scikit_learn==1.6.1
torch==2.5.1
torchvision==0.20.0
tqdm==4.67.0
umap_learn==0.5.7
